feat: add professional tooling — ruff, mypy, pre-commit, dependabot, coverage
- Add pyproject.toml consolidating pytest, ruff, mypy, coverage config
- Add ruff lint + format: fix 118 violations (unused imports, isort, etc.)
- Add mypy type checking with pragmatic ignores for PyQt6/Win32
- Add pre-commit hooks (ruff, trailing whitespace, detect-private-key)
- Add dependabot for pip + github-actions dependency updates
- Add codecov coverage reporting to CI
- Add lint.yml and typecheck.yml CI workflows
- Fix type annotations: implicit Optional, missing type hints
- Remove pytest.ini (consolidated into pyproject.toml)

+++ gui/widgets.py
@@ -1,8 +1,7 @@
 """Custom HUD widgets for Jarvis GUI."""
 
-import math
-from PyQt6.QtCore import Qt, QTimer, QRectF, QPointF
-from PyQt6.QtGui import QPainter, QColor, QPen, QBrush, QRadialGradient, QFont
+from PyQt6.QtCore import QPointF, QRectF, Qt, QTimer
+from PyQt6.QtGui import QBrush, QColor, QFont, QPainter, QPen, QRadialGradient
 from PyQt6.QtWidgets import QWidget
 
 from gui.styles import PRIMARY_COLOR, SECONDARY_COLOR, TEXT_COLOR
@@ -66,10 +65,10 @@ class ArcReactor(QWidget):
         painter.setPen(arc_pen)
 
         rect = QRectF(cx - radius + 15, cy - radius + 15, (radius - 15) * 2, (radius - 15) * 2)
-        painter.drawArc(rect, int(self._angle * 16), int(90 * 16))
+        painter.drawArc(rect, int(self._angle * 16), (90 * 16))
 
         rect2 = QRectF(cx - radius + 25, cy - radius + 25, (radius - 25) * 2, (radius - 25) * 2)
-        painter.drawArc(rect2, int(-self._angle * 16 + 180 * 16), int(60 * 16))
+        painter.drawArc(rect2, int(-self._angle * 16 + 180 * 16), (60 * 16))
 
         # Inner circle
         inner_radius = radius * 0.4
